feat: refactor to support versioning of api/v1
